Refactor: Replace Team and Admin pages with new Dashboard page
