+++ app/globals.css
@@ -40,15 +40,15 @@
     @apply m-auto flex flex-1 flex-col items-center justify-between gap-10 py-10;
   }
 
-  .admin-header, .team-header {
+  .dashboard-header {
     @apply sticky top-3 z-20 mx-3 flex items-center justify-center rounded-2xl bg-dark-200 px-[5%] py-5 shadow-lg xl:px-12;
   }
 
-  .admin-main {
+  .dashboard-main {
     @apply flex flex-col items-center space-y-6 px-[5%] pb-12 xl:space-y-12 xl:px-12;
   }
 
-  .admin-stat {
+  .dashboard-stat {
     @apply flex w-full flex-col justify-between gap-5 sm:flex-row xl:gap-10;
   }
