+++ app/login/page.tsx
@@ -8,12 +8,10 @@ import { z } from "zod";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { PasswordSchema } from "../../lib/validation";
 import CustomFormField, { FormFieldType } from "../../components/commons/CustomFormField";
-import { RadioGroup } from "@radix-ui/react-radio-group";
 
 const LoginPage = () => {
   const router = useRouter();
   const { login } = useAuth();
-  const [destination, setDestination] = useState("admin");
   const [errorMessage, setErrorMessage] = useState("");
 
   const form = useForm<z.infer<typeof PasswordSchema>>({
@@ -30,12 +28,7 @@ const LoginPage = () => {
 
       if (data.userId) {
         login(data.userId);
-
-        if (destination === "admin") {
-          router.push(`/mediaExperts/${data.userId}/admin`);
-        } else {
-          router.push(`/mediaExperts/${data.userId}/team`);
-        }
+        router.push(`/mediaExperts/${data.userId}/dashboard`);
       } else {
         setErrorMessage("Invalid password");
       }
@@ -67,30 +60,6 @@ const LoginPage = () => {
                 maxLength={50}
                 minLength={6}
               />
-            </div>
-            <div className="flex justify-center mb-6">
-              <RadioGroup className="flex justify-center space-x-4 mb-6 radio-group">
-                <label className="flex items-center space-x-2">
-                  <input
-                    type="radio"
-                    value="admin"
-                    checked={destination === "admin"}
-                    onChange={() => setDestination("admin")}
-                    className="login-radio"
-                  />
-                  <span>Admin</span>
-                </label>
-                <label className="flex items-center space-x-2 ml-4">
-                  <input
-                    type="radio"
-                    value="team"
-                    checked={destination === "team"}
-                    onChange={() => setDestination("team")}
-                    className="login-radio"
-                  />
-                  <span>Team</span>
-                </label>
-              </RadioGroup>
             </div>
             <div className="flex justify-center">
               <button type="submit" className="login-radio-button">
