+++ app/mediaExperts/[userId]/courses/[courseId]/addParticipant/page.tsx
@@ -6,7 +6,7 @@ const AddParticipantPage = ({ params, searchParams }: { params: { userId: string
     title="Participant Added Successfully"
     description="Participant has been successfully added."
     buttonText="Go to Dashboard"
-    buttonLink={`/mediaExperts/${params.userId}/team`}
+    buttonLink={`/mediaExperts/${params.userId}/dashboard`}
   />
   );
 };
