+++ app/mediaExperts/[userId]/courses/[courseId]/joinCourse/page.tsx
@@ -6,7 +6,7 @@ const JoinCoursePage = ({ params }: { params: { userId: string } }) => {
       title="Course Joined Successfully"
       description="You have successfully joined the course."
       buttonText="Go to Dashboard"
-      buttonLink={`/mediaExperts/${params.userId}/admin`}
+      buttonLink={`/mediaExperts/${params.userId}/dashboard`}
     />
   );
   }
