--- app/mediaExperts/[userId]/team/page.tsx
@@ -1,73 +1,0 @@
-"use client";
-
-import { useEffect, useState } from "react";
-import { columnsCourse } from "../../../../components/table/columnsCourse";
-import { columnsPlatform } from "../../../../components/table/columnsPlatform";
-import { DataTable } from "../../../../components/table/DataTable";
-import { getCourses } from "../../../../lib/actions/course.actions";
-import { getMediaExperts } from "../../../../lib/actions/media_expert.actions";
-import { getPlatforms } from "../../../../lib/actions/platform.actions";
-import { ICourse, IPlatform } from "../../../../types/appwrite.types";
-import AddParticipant from "../../../../components/AddParticipant";
-import Course from "../../../../components/Course";
-import Platform from "../../../../components/Platform";
-import PublishContent from "../../../../components/PublishContent";
-
-const TeamPage = ({ params }: { params: { userId: string } }) => {
-  const { userId } = params;
-  const [coursesData, setCoursesData] = useState<ICourse[]>([]);
-  const [mediaExperts, setMediaExperts] = useState<IMediaExpert[]>([]);
-  const [platformsData, setPlatformsData] = useState<IPlatform[]>([]);
-  const [loading, setLoading] = useState<boolean>(true);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const [courses, mediaExperts, platforms] = await Promise.all([
-          getCourses(),
-          getMediaExperts(),
-          getPlatforms(),
-        ]);
-        setCoursesData(courses);
-        setMediaExperts(mediaExperts);
-        setPlatformsData(platforms);
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-    fetchData();
-  }, []);
-
-  if (loading) {
-    return <div>Loading...</div>;
-  }
-
-  return (
-    <div className="mx-auto flex max-w-7xl flex-col space-y-14">
-      <header className="team-header mt-8">
-        <p className="text-16-bold text-center">Team Dashboard</p>
-      </header>
-
-      <main className="admin-main">
-        <section className="w-full flex flex-col md:flex-row justify-around md:space-x-4 space-y-4 md:space-y-0">
-          <Course userId={userId} closeModal={() => {}} />
-          <AddParticipant userId={userId} />
-        </section>
-        <section className="w-full space-y-4">
-          <DataTable columns={columnsCourse} data={coursesData} />
-        </section>
-        <section className="w-full flex flex-col md:flex-row justify-around md:space-x-4 space-y-4 md:space-y-0">
-          <Platform userId={userId} closeModal={() => {}} />
-          <PublishContent userId={userId} />
-        </section>
-        <section className="w-full space-y-4">
-          <DataTable columns={columnsPlatform} data={platformsData} />
-        </section>
-      </main>
-    </div>
-  );
-};
-
-export default TeamPage;
